requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
websockets==11.0.3
python-dotenv==1.0.0
requests==2.31.0
chromadb==0.4.18
beautifulsoup4==4.12.2
PyYAML==6.0.1

# UI dependencies
jinja2==3.1.2

# Computer Vision Assistant dependencies
selenium==4.18.1
webdriver-manager==4.0.1
pyautogui==0.9.54
psutil==5.9.6
rich==13.7.0
click==8.1.7

# Development dependencies
pytest==7.4.3
black==23.10.1
isort==5.12.0
flake8==6.1.0
